Move code to apps.ts
