+++ src/commands/index.js
@@ -59,8 +59,6 @@ module.exports = function(client) {
   client.login.ci = loadCommand("login-ci");
   client.logout = loadCommand("logout");
   client.open = loadCommand("open");
-  client.projects = {};
-  client.projects.list = loadCommand("projects-list");
   client.serve = loadCommand("serve");
   client.setup = {};
   client.setup.emulators = {};
