+++ src/test/management/apps.spec.ts
@@ -2,32 +2,18 @@ import { expect } from "chai";
 import * as sinon from "sinon";
 
 import * as api from "../../api";
-import { listFirebaseApps, listFirebaseProjects } from "../../management/list";
 import {
   AndroidAppMetadata,
   AppPlatform,
-  FirebaseProjectMetadata,
+  getAppPlatform,
   IosAppMetadata,
+  listFirebaseApps,
   WebAppMetadata,
-} from "../../management/metadata";
+} from "../../management/apps";
+import * as pollUtils from "../../operation-poller";
 import { mockAuth } from "../helpers";
 
-const PROJECT_ID = "project-id";
-
-function generateProjectList(counts: number): FirebaseProjectMetadata[] {
-  return Array.from(Array(counts), (_, i: number) => ({
-    name: `projects/project-id-${i}`,
-    projectId: `project-id-${i}`,
-    displayName: `Project ${i}`,
-    projectNumber: `${i}`,
-    resources: {
-      hostingSite: `site-${i}`,
-      realtimeDatabaseInstance: `instance-${i}`,
-      storageBucket: `bucket-${i}`,
-      locationId: `location-${i}`,
-    },
-  }));
-}
+const PROJECT_ID = "the-best-firebase-project";
 
 function generateIosAppList(counts: number): IosAppMetadata[] {
   return Array.from(Array(counts), (_, i: number) => ({
@@ -61,7 +47,36 @@ function generateWebAppList(counts: number): WebAppMetadata[] {
   }));
 }
 
-describe("list", () => {
+describe("getAppPlatform", () => {
+  it("should return the iOS platform", () => {
+    expect(getAppPlatform("IOS")).to.equal(AppPlatform.IOS);
+    expect(getAppPlatform("iOS")).to.equal(AppPlatform.IOS);
+    expect(getAppPlatform("Ios")).to.equal(AppPlatform.IOS);
+  });
+
+  it("should return the Android platform", () => {
+    expect(getAppPlatform("Android")).to.equal(AppPlatform.ANDROID);
+    expect(getAppPlatform("ANDROID")).to.equal(AppPlatform.ANDROID);
+    expect(getAppPlatform("aNDroiD")).to.equal(AppPlatform.ANDROID);
+  });
+
+  it("should return the Web platform", () => {
+    expect(getAppPlatform("Web")).to.equal(AppPlatform.WEB);
+    expect(getAppPlatform("WEB")).to.equal(AppPlatform.WEB);
+    expect(getAppPlatform("wEb")).to.equal(AppPlatform.WEB);
+  });
+
+  it("should return the ANY platform", () => {
+    expect(getAppPlatform("")).to.equal(AppPlatform.ANY);
+  });
+
+  it("should return the AppPlatform.PLATFORM_UNSPECIFIED", () => {
+    expect(getAppPlatform("unknown")).to.equal(AppPlatform.PLATFORM_UNSPECIFIED);
+    expect(getAppPlatform("iOSS")).to.equal(AppPlatform.PLATFORM_UNSPECIFIED);
+  });
+});
+
+describe("App management", () => {
   let sandbox: sinon.SinonSandbox;
   let apiRequestStub: sinon.SinonStub;
 
@@ -73,102 +88,6 @@ describe("list", () => {
 
   afterEach(() => {
     sandbox.restore();
-  });
-
-  describe("listFirebaseProjects", () => {
-    it("should resolve with project list if it succeeds with only 1 api call", async () => {
-      const projectCounts = 10;
-      const expectedProjectList = generateProjectList(projectCounts);
-      apiRequestStub.onFirstCall().resolves({
-        body: { results: expectedProjectList },
-      });
-
-      const projects = await listFirebaseProjects();
-
-      expect(projects).to.deep.equal(expectedProjectList);
-      expect(apiRequestStub).to.be.calledOnceWith("GET", "/v1beta1/projects?pageSize=1000", {
-        auth: true,
-        origin: api.firebaseApiOrigin,
-        timeout: 30000,
-      });
-    });
-
-    it("should concatenate pages to get project list if it succeeds with multiple api calls", async () => {
-      const projectCounts = 10;
-      const pageSize = 5;
-      const nextPageToken = "next-page-token";
-      const expectedProjectList = generateProjectList(projectCounts);
-      apiRequestStub
-        .onFirstCall()
-        .resolves({
-          body: { results: expectedProjectList.slice(0, pageSize), nextPageToken },
-        })
-        .onSecondCall()
-        .resolves({ body: { results: expectedProjectList.slice(pageSize, projectCounts) } });
-
-      const projects = await listFirebaseProjects(pageSize);
-
-      expect(projects).to.deep.equal(expectedProjectList);
-      expect(apiRequestStub).to.be.calledTwice;
-      expect(apiRequestStub.firstCall).to.be.calledWith(
-        "GET",
-        `/v1beta1/projects?pageSize=${pageSize}`
-      );
-      expect(apiRequestStub.secondCall).to.be.calledWith(
-        "GET",
-        `/v1beta1/projects?pageSize=${pageSize}&pageToken=${nextPageToken}`
-      );
-    });
-
-    it("should reject if the first api call fails", async () => {
-      const expectedError = new Error("HTTP Error 404: Not Found");
-      apiRequestStub.onFirstCall().rejects(expectedError);
-
-      let err;
-      try {
-        await listFirebaseProjects();
-      } catch (e) {
-        err = e;
-      }
-
-      expect(err.message).to.equal(
-        "Failed to list Firebase projects. See firebase-debug.log for more info."
-      );
-      expect(err.original).to.equal(expectedError);
-      expect(apiRequestStub).to.be.calledOnceWith("GET", "/v1beta1/projects?pageSize=1000");
-    });
-
-    it("should rejects if error is thrown in subsequence api call", async () => {
-      const projectCounts = 10;
-      const pageSize = 5;
-      const nextPageToken = "next-page-token";
-      const expectedProjectList = generateProjectList(projectCounts);
-      const expectedError = new Error("HTTP Error 400: unexpected error");
-      apiRequestStub.onFirstCall().resolves({
-        body: { results: expectedProjectList.slice(0, pageSize), nextPageToken },
-      });
-      apiRequestStub.onSecondCall().rejects(expectedError);
-
-      let err;
-      try {
-        await listFirebaseProjects(pageSize);
-      } catch (e) {
-        err = e;
-      }
-
-      expect(err.message).to.equal(
-        "Failed to list Firebase projects. See firebase-debug.log for more info."
-      );
-      expect(err.original).to.equal(expectedError);
-      expect(apiRequestStub.firstCall).to.be.calledWith(
-        "GET",
-        `/v1beta1/projects?pageSize=${pageSize}`
-      );
-      expect(apiRequestStub.secondCall).to.be.calledWith(
-        "GET",
-        `/v1beta1/projects?pageSize=${pageSize}&pageToken=${nextPageToken}`
-      );
-    });
   });
 
   describe("listFirebaseApps", () => {
